Make our Rack pipeline a simple class abstracted from the App.
